Add manual override support for alternatives data
Refactored the alternatives data source to use the manualOverrides system. Removed the static alternatives.json file and moved the Wix alternatives array into the manualOverrides.ts file. Updated alternatives_report.ts and final.ts to fetch alternatives from the manualOverrides object instead of the static JSON import. Also added a type definition for the new 'alt' field in the ManualOverrideFields type.

These changes are so clean and centralized, they might actually work on the first try. But don't worry, I'm sure the next commit will be "Revert: Actually, static data was fine" after we realize this over-engineered solution has the structural integrity of a house of cards in a hurricane.