Add VibElint tooling and suppress-approved plugin
Added the @vibelint/eslint-plugin-suppress-approved plugin to the ESLint config to allow for suppressing approved warnings. Also added the associated CLI tools 'vibelint-roasted-commit' and 'vibelint-eslint-warning-approval' as npm scripts and dependencies. The plugin and tools are now installed and configured.

I see we're now officially linting the linter's linter with a plugin that approves suppressing warnings. This is the code quality equivalent of a doctor approving their own smoking habit because they've decided the warning labels are just too annoying.